Create util
